[BE] ✨ 마켓 정보 리팩토링 및 무한 스크롤 페이징 구현
